Aquabid script
This script allows you to click on the empty lines at aquabid and then makes an ajax request to get the link and view only the images on the following page.  This means that instead of having to click and go to the page you can just see the images of the page.  It actually makes browsing aquabid faster.